Function sui_transactional_test_runner::create_adapter

pub async fn create_adapter<'a, Adapter>(
    path: &Path,
    fully_compiled_program_opt: Option<Arc<FullyCompiledProgram>>,
) -> Result<(String, Adapter), Box<dyn Error>>
where
    Adapter: MoveTestAdapter<'a>,
    <Adapter as MoveTestAdapter<'a>>::ExtraInitArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::ExtraPublishArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::ExtraValueArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::ExtraRunArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::Subcommand: Debug,

Creates an adapter for the given tasks, using the first task command to initialize the adapter if it is a TaskCommand::Init. Returns the adapter and the output string.