sui_transactional_test_runner

Function run_test_impl

pub async fn run_test_impl<'a, Adapter>(
    path: &Path,
    pre_compiled_program: Option<Arc<PreCompiledProgramInfo>>,
    insta_options: Option<InstaOptions>,
) -> Result<(), Box<dyn Error>>
where
    Adapter: MoveTestAdapter<'a>,
    <Adapter as MoveTestAdapter<'a>>::ExtraInitArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::ExtraPublishArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::ExtraValueArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::ExtraRunArgs: Debug,
    <Adapter as MoveTestAdapter<'a>>::Subcommand: Debug,

Convenience function that creates an adapter and runs the tasks, to be used when a caller does not need to extend the adapter.