sui_transactional_test_runner::offchain_state
Trait OffchainStateReader

pub trait OffchainStateReader:
    Send
    + Sync
    + 'static {
    // Required methods
    fn wait_for_checkpoint_catchup<'life0, 'async_trait>(
        &'life0 self,
        checkpoint: u64,
        base_timeout: Duration,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wait_for_pruned_checkpoint<'life0, 'async_trait>(
        &'life0 self,
        checkpoint: u64,
        base_timeout: Duration,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute_graphql<'life0, 'async_trait>(
        &'life0 self,
        query: String,
        show_usage: bool,
    ) -> Pin<Box<dyn Future<Output = Result<TestResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute_jsonrpc<'life0, 'async_trait>(
        &'life0 self,
        method: String,
        params: Value,
    ) -> Pin<Box<dyn Future<Output = Result<TestResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Trait for interacting with the offchain state of the Sui network. To reduce test flakiness, these methods are used in the RunGraphqlCommand to stabilize the off-chain indexed state.

Required Methods

fn wait_for_checkpoint_catchup<'life0, 'async_trait>( &'life0 self, checkpoint: u64, base_timeout: Duration, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Polls the checkpoint table until the given checkpoint is committed.

fn wait_for_pruned_checkpoint<'life0, 'async_trait>( &'life0 self, checkpoint: u64, base_timeout: Duration, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Polls the checkpoint table until the given checkpoint is pruned.

fn execute_graphql<'life0, 'async_trait>( &'life0 self, query: String, show_usage: bool, ) -> Pin<Box<dyn Future<Output = Result<TestResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Executes a GraphQL query and returns the response.

fn execute_jsonrpc<'life0, 'async_trait>( &'life0 self, method: String, params: Value, ) -> Pin<Box<dyn Future<Output = Result<TestResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Executes a JSON-RPC query and returns the response. The query is given as a method name, and a JSON value representing its parameters.