sui_transactional_test_runner::test_adapter
Struct SuiTestAdapter

pub struct SuiTestAdapter {
    pub read_replica: Option<Arc<dyn RpcStateReader + Send + Sync>>,
    pub offchain_config: Option<OffChainConfig>,
    pub offchain_reader: Option<Box<dyn OffchainStateReader>>,
    /* private fields */
}

Fields
read_replica: Option<Arc<dyn RpcStateReader + Send + Sync>>

If is_simulator is true, the executor will be a Simulacrum, and this will be a RpcStateReader that can be used to spawn the equivalent of a fullnode rest api. This can then be used to serve an indexer that reads from said rest api service.

offchain_config: Option<OffChainConfig>

Configuration for offchain state reader read from the file itself, and can be passed to the specific indexing and reader flavor.

offchain_reader: Option<Box<dyn OffchainStateReader>>

A trait encapsulating methods to interact with offchain state.

Implementations

impl SuiTestAdapter

pub fn with_offchain_reader( &mut self, offchain_reader: Box<dyn OffchainStateReader>, )

pub fn is_simulator(&self) -> bool

pub fn executor(&self) -> &dyn TransactionalAdapter

pub fn into_executor(self) -> Box<dyn TransactionalAdapter>