Crate sui_transactional_test_runner

This module contains the transactional test runner instantiation for the Sui adapter

Modules

args

programmable_transaction_test_parser

test_adapter

This module contains the transactional test runner instantiation for the Sui adapter

Structs

ValidatorWithFullnode

Traits

TransactionalAdapter

TODO: better name?

Functions

run_test

run_test_impl