Crate sui_transactional_test_runner

This module contains the transactional test runner instantiation for the Sui adapter

Modules

args

offchain_state

programmable_transaction_test_parser

test_adapter

This module contains the transactional test runner instantiation for the Sui adapter

Structs

ValidatorWithFullnode

Traits

TransactionalAdapter

TODO: better name?

Functions

create_adapter

Creates an adapter for the given tasks, using the first task command to initialize the adapter if it is a TaskCommand::Init. Returns the adapter and the output string.

run_tasks_with_adapter

Consumes the adapter to run tasks from path.

run_test

run_test_impl

Convenience function that creates an adapter and runs the tasks, to be used when a caller does not need to extend the adapter.